Add files via upload
In this project, created a classification machine learning model that can predict whether patients are having a seizure or not through EEG readings. The classifiers that were used for this study were Naïve Bayes and Logistic Regression. Using the PCA method to minimized the components of the experiment and suggested the most optimal number of components in order to succeed the highest accuracy.